requirements.txt
Pillow
opencv-python
python-dotenv
mysqlclient
numpy
fastapi
python-multipart
uvicorn
torch
torchvision